dev-6533 integrate api to differences container
